move color parsing function to pysynap/utils.py
