--- pysynap/utils/colors_coco.json
@@ -1,82 +1,0 @@
-{
-    "0": [189, 114, 0],
-    "1": [25, 83, 217],
-    "2": [32, 177, 237],
-    "3": [142, 47, 126],
-    "4": [48, 172, 119],
-    "5": [238, 190, 77],
-    "6": [47, 20, 162],
-    "7": [77, 77, 77],
-    "8": [153, 153, 153],
-    "9": [0, 0, 255],
-    "10": [0, 128, 255],
-    "11": [0, 191, 191],
-    "12": [0, 255, 0],
-    "13": [255, 0, 0],
-    "14": [255, 0, 170],
-    "15": [0, 85, 85],
-    "16": [0, 170, 85],
-    "17": [0, 255, 85],
-    "18": [0, 85, 170],
-    "19": [0, 170, 170],
-    "20": [0, 255, 170],
-    "21": [0, 85, 255],
-    "22": [0, 170, 255],
-    "23": [0, 255, 255],
-    "24": [128, 85, 0],
-    "25": [128, 170, 0],
-    "26": [128, 255, 0],
-    "27": [85, 0, 128],
-    "28": [85, 85, 128],
-    "29": [85, 170, 128],
-    "30": [85, 255, 128],
-    "31": [170, 0, 128],
-    "32": [170, 85, 128],
-    "33": [170, 170, 128],
-    "34": [170, 255, 128],
-    "35": [255, 0, 128],
-    "36": [255, 85, 128],
-    "37": [255, 170, 128],
-    "38": [255, 255, 128],
-    "39": [255, 85, 0],
-    "40": [255, 170, 0],
-    "41": [255, 255, 0],
-    "42": [128, 0, 85],
-    "43": [128, 85, 85],
-    "44": [128, 170, 85],
-    "45": [128, 255, 85],
-    "46": [170, 0, 85],
-    "47": [170, 85, 85],
-    "48": [170, 170, 85],
-    "49": [170, 255, 85],
-    "50": [255, 0, 85],
-    "51": [255, 85, 85],
-    "52": [255, 170, 85],
-    "53": [255, 255, 255],
-    "54": [0, 0, 85],
-    "55": [0, 0, 128],
-    "56": [0, 0, 170],
-    "57": [0, 0, 212],
-    "58": [0, 0, 255],
-    "59": [0, 85, 0],
-    "60": [0, 128, 0],
-    "61": [0, 170, 0],
-    "62": [0, 212, 0],
-    "63": [0, 255, 0],
-    "64": [85, 0, 0],
-    "65": [128, 0, 0],
-    "66": [170, 0, 0],
-    "67": [212, 0, 0],
-    "68": [255, 0, 0],
-    "69": [85, 85, 0],
-    "70": [0, 85, 128],
-    "71": [0, 85, 170],
-    "72": [0, 85, 212],
-    "73": [0, 85, 255],
-    "74": [0, 128, 85],
-    "75": [0, 128, 128],
-    "76": [0, 128, 170],
-    "77": [0, 128, 212],
-    "78": [0, 128, 255],
-    "79": [0, 170, 85]
-}
